% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodepolyline.R
\name{decodepolyline}
\alias{decodepolyline}
\title{Decode an encoded polyline}
\usage{
decodepolyline(encoded)
}
\arguments{
\item{encoded}{is an encoded polyline}
}
\value{
Returns a data frame
}
\description{
Decode an encoded polyline
}
\examples{

# Request compressed polylines for each route:
mymultiroute <- multiroute(fromcoord = "55.68958,12.557845", tocoords = coords, routepolyline = 1)

# Decode the compressed polyline
decodepolyline(mymultiroute[[1]]$routepolyline)

}

