% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apicalls.R
\name{POInear}
\alias{POInear}
\title{Nearest POI of Given Types}
\usage{
POInear(fromaddress = NULL, fromcoord = NULL, poitypes = NULL,
  mot = NULL, routepolyline = NULL)
}
\arguments{
\item{fromaddress}{is an adress type: roadname+housenumber+postalcode}

\item{fromcoord}{is a coordinate type: latitude, longtitude}

\item{poitypes}{is a vector poi types to search for}

\item{mot}{is a mode of transport designator
\itemize{
  \item car (default)
  \item bicycle
  \item foot
  \item public
  \item auto
}}

\item{routepolyline}{append a route in the form of a list of coordinates.
\itemize{
  \item 0 (default): without route description
  \item 1: with route description as a an encoded polyline string
  \item 2: with route description as a list of latitude,longtitude pairs.
}}
}
\value{
Returns a list
}
\description{
Returns the POI for each requested type that has the shortes route to the outset point.
}
\examples{

# Request with address as origintype:
POInear(fromaddress = "elmegade 5 2200",
        poitypes = c("daycare", "school", "supermarket", "hospital", "doctor", "train", "stop", "junction", "metro", "strain")
)

# Request with coordinate as origintype:
POInear(fromcoord = "56.16390607394954,10.16510009765625",
        poitypes = c("daycare", "school", "supermarket", "hospital", "doctor", "train", "stop", "junction", "metro", "strain")
)

# Set mode of transport to bicycle:
POInear(fromcoord = "56.16390607394954,10.16510009765625",
        poitypes = c("daycare", "school", "supermarket", "hospital", "doctor", "train", "stop", "junction", "metro", "strain"),
        mot = "bicycle"
)

# Request compressed polylines for each of the nearest POIs:
POInear(fromcoord = "56.16390607394954,10.16510009765625",
        poitypes = c("daycare", "school", "supermarket", "hospital", "doctor", "train", "stop", "junction", "metro", "strain"),
        routepolyline = 1
)

}

