% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apicalls.R
\name{POIradius}
\alias{POIradius}
\title{Points of Interest Within A Given Radius}
\usage{
POIradius(fromaddress = NULL, fromcoord = NULL, poitypes = NULL)
}
\arguments{
\item{fromaddress}{is an adress type: roadname+housenumber+postalcode}

\item{fromcoord}{is a coordinate type: latitude, longtitude}

\item{poitypes}{is a vector poi types to search for}
}
\value{
Returns a list
}
\description{
Returns a list of name and coordinates of POI's based on the search parameters.
}
\examples{

# POIs of type pharmacy, lake or coast within default 20km radius of address:
POIradius(fromaddress = "elmegade 5 2200", poitypes = c("pharmacy", "lake", "coast", "forest"))

# POIs of type pharmacy, lake or coast within default 20km radius of coordinate:
POIradius(fromcoord = "55.69043,12.53527", poitypes = c("pharmacy", "lake", "coast", "forest"))

}

