% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apicalls.R
\name{POIwithin}
\alias{POIwithin}
\title{Points of Interest Within A Geometry}
\usage{
POIwithin(poitypes = NULL, bbox = NULL, municipalcode = NULL)
}
\arguments{
\item{poitypes}{is a vector poi types to search for.}

\item{bbox}{is a boundingbox: lower_lat, lower_lng, upper_lat, upper_lng. Must be a data frame with a lat and lng column.}

\item{municipalcode}{municipalcode: one of 98 danish municipality codes}
}
\value{
Returns a list
}
\description{
Returns a list of name and coordinates of POI's based on the search parameters.
}
\examples{

# Make bounding box coordinates
coords <- dplyr::data_frame(
  lat = c(54.842508, 56.131587),
  lng = c(9.633364, 12.574262))

# POIs of type airport within a large bounding box:
POIwithin(poitypes = "airport", bbox = coords)

# Libraries within municiaplity of Copenhagen:
POIwithin(poitypes = "library", municipalcode = 101)

}

