% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apicalls.R
\name{tsp}
\alias{tsp}
\title{Travelling Salesperson}
\usage{
tsp(fromaddress = NULL, fromcoord = NULL, tocoords = NULL,
  metric = NULL, zvalue = NULL)
}
\arguments{
\item{fromaddress}{is an adress type: roadname+housenumber+postalcode}

\item{fromcoord}{is a coordinate type: latitude, longtitude}

\item{tocoords}{is a coordinate type: latitude, longtitude. Must be a data frame with a lat and lng column}

\item{metric}{is a choice of cost metric for optimal route. Either "time" (default) or "distance"}

\item{zvalue}{is the zoom level for polyline. Integer in the range 1 to 19. Default: 18}
}
\value{
Returns a list
}
\description{
Calculate shortest route visiting every destination before returning home.
}
\examples{

# Create data frame with coordinates
coords <- dplyr::data_frame(
  lat = c(55.70873874051673, 55.68745586048921, 55.67274455523506,
          55.70390273988158, 55.62799595426723, 55.7425739894847),
  lng = c(12.581062316894531, 12.496261596679688, 12.48870849609375,
          12.363052368164062, 12.55256652832031, 12.454376220703125))

# TSP optimized for shortest travelling time:
tsp(fromcoord = "55.6896,12.55792", tocoords = coords, metric = "time")

# TSP optimized for shortest travelling distance:
tsp(fromcoord = "55.6896,12.55792", tocoords = coords, metric = "distance")

# Decode the compressed polyline
mytsp <- tsp(fromcoord = "55.6896,12.55792", tocoords = coords, metric = "distance")

decodepolyline(mytsp$steps[[1]]$route_geometry)

}

